requirements.txt
pandas>=1.3
pyyaml
requests
python-dotenv
numpy
